data_processor/requirements.txt
pika==1.2.0
psycopg2-binary==2.9.3
pydantic==2.5.2
python-dateutil==2.8.2
